All flow done

+++ MercadoLibre/app/src/main/java/com/andres/mercadolibre/api/core/model/PayerCostModel.java
@@ -9,13 +9,13 @@ public class PayerCostModel {
 
     @SerializedName("installments")
     @Expose
-    public Integer installments;
+    public Float installments;
     @SerializedName("installment_rate")
     @Expose
     public Double installmentRate;
     @SerializedName("discount_rate")
     @Expose
-    public Integer discountRate;
+    public Float discountRate;
     @SerializedName("labels")
     @Expose
     public List<String> labels = null;
@@ -24,18 +24,18 @@ public class PayerCostModel {
     public List<String> installmentRateCollector = null;
     @SerializedName("min_allowed_amount")
     @Expose
-    public Integer minAllowedAmount;
+    public Float minAllowedAmount;
     @SerializedName("max_allowed_amount")
     @Expose
-    public Integer maxAllowedAmount;
+    public Float maxAllowedAmount;
     @SerializedName("recommended_message")
     @Expose
     public String recommendedMessage;
     @SerializedName("installment_amount")
     @Expose
-    public Integer installmentAmount;
+    public Float installmentAmount;
     @SerializedName("total_amount")
     @Expose
-    public Integer totalAmount;
+    public Float totalAmount;
 
 }
